SSL: attempt to squelch a narrowing warning.
SSL_set_fd() takes an int as its second argument; that's not an issue on
UN*X, as sockets are represented by file descriptors, but causes
warnings on Windows, where a SOCKET is not an int.

A comment in OpenSSL's intternal sockets.h header says, on 64-bit
Windows:

	Even though sizeof(SOCKET) is 8, it's safe to cast it to int,
	because the value constitutes an index in per-process table of
	limited size and not a real pointer.

so the cast should be safe.

+++ sslutils.c
@@ -140,7 +140,7 @@ SSL *ssl_promotion(int is_server, SOCKET s, char *errbuf, size_t errbuflen)
 	}
 
 	SSL *ssl = SSL_new(ctx); // TODO: also a DTLS context
-	SSL_set_fd(ssl, s);
+	SSL_set_fd(ssl, (int)s);
 
 	if (is_server) {
 		if (SSL_accept(ssl) <= 0) {
